Move the rt::io::process tests to run-pass
Closes #9341
